#1190: refactorings and additions moved old installPlugin method (used by android-sudio) into new IdeaPluginDownloader class moved installPlugin method of intellij to IdeaBasedToolCommandlet class added help entries for pycharm added changelog entry added missing binary files for tests
