+++ cli/src/main/java/com/devonfw/tools/ide/tool/androidstudio/AndroidStudio.java
@@ -6,9 +6,13 @@ import java.util.Set;
 import com.devonfw.tools.ide.common.Tag;
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.process.EnvironmentContext;
+import com.devonfw.tools.ide.process.ProcessContext;
+import com.devonfw.tools.ide.step.Step;
 import com.devonfw.tools.ide.tool.ToolInstallation;
 import com.devonfw.tools.ide.tool.ide.IdeToolCommandlet;
 import com.devonfw.tools.ide.tool.ide.IdeaBasedIdeToolCommandlet;
+import com.devonfw.tools.ide.tool.ide.IdeaPluginDownloader;
+import com.devonfw.tools.ide.tool.plugin.ToolPluginDescriptor;
 
 /**
  * {@link IdeToolCommandlet} for <a href="https://developer.android.com/studio">AndroidStudio</a>.
@@ -52,5 +56,11 @@ public class AndroidStudio extends IdeaBasedIdeToolCommandlet {
 
     super.setEnvironment(environmentContext, toolInstallation, extraInstallation);
     environmentContext.withEnvVar("STUDIO_PROPERTIES", this.context.getWorkspacePath().resolve("studio.properties").toString());
+  }
+
+  @Override
+  public boolean installPlugin(ToolPluginDescriptor plugin, Step step, ProcessContext pc) {
+    IdeaPluginDownloader ideaPluginDownloader = new IdeaPluginDownloader(this.context, this);
+    return ideaPluginDownloader.installPlugin(plugin, step, pc);
   }
 }
