+++ cli/src/main/java/com/devonfw/tools/ide/tool/intellij/Intellij.java
@@ -1,22 +1,14 @@
 package com.devonfw.tools.ide.tool.intellij;
 
 import java.nio.file.Files;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Set;
 
 import com.devonfw.tools.ide.common.Tag;
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.process.EnvironmentContext;
-import com.devonfw.tools.ide.process.ProcessContext;
-import com.devonfw.tools.ide.process.ProcessErrorHandling;
-import com.devonfw.tools.ide.process.ProcessMode;
-import com.devonfw.tools.ide.process.ProcessResult;
-import com.devonfw.tools.ide.step.Step;
 import com.devonfw.tools.ide.tool.ToolInstallation;
 import com.devonfw.tools.ide.tool.ide.IdeToolCommandlet;
 import com.devonfw.tools.ide.tool.ide.IdeaBasedIdeToolCommandlet;
-import com.devonfw.tools.ide.tool.plugin.ToolPluginDescriptor;
 
 /**
  * {@link IdeToolCommandlet} for <a href="https://www.jetbrains.com/idea/">IntelliJ</a>.
@@ -60,28 +52,6 @@ public class Intellij extends IdeaBasedIdeToolCommandlet {
 
     super.setEnvironment(environmentContext, toolInstallation, extraInstallation);
     environmentContext.withEnvVar("IDEA_PROPERTIES", this.context.getWorkspacePath().resolve("idea.properties").toString());
-  }
-
-  @Override
-  public boolean installPlugin(ToolPluginDescriptor plugin, final Step step, ProcessContext pc) {
-
-    // In case of plugins with a custom repo url
-    boolean customRepo = plugin.url() != null;
-    List<String> args = new ArrayList<>();
-    args.add("installPlugins");
-    args.add(plugin.id());
-    if (customRepo) {
-      args.add(plugin.url());
-    }
-    ProcessResult result = runTool(ProcessMode.DEFAULT, ProcessErrorHandling.LOG_WARNING, pc, args.toArray(String[]::new));
-    if (result.isSuccessful()) {
-      this.context.success("Successfully installed plugin: {}", plugin.name());
-      step.success();
-      return true;
-    } else {
-      step.error("Failed to install plugin {} ({}): exit code was {}", plugin.name(), plugin.id(), result.getExitCode());
-      return false;
-    }
   }
 
 }
