+++ cli/src/main/resources/nls/Help.properties
@@ -78,6 +78,8 @@ cmd.oc=Tool commandlet for Openshift CLI (Kubernetes management tool).
 cmd.oc.detail=OpenShift CLI (oc) is a command-line tool for managing OpenShift clusters. Detailed documentation can be found at https://docs.openshift.com/
 cmd.pgadmin=Tool commandlet for pgAdmin.
 cmd.pgadmin.detail=pgAdmin is an open source management tool for PostgreSQL and derivative relational databases. Detailed documentation can be found at https://www.pgadmin.org/docs/pgadmin4/latest/index.html
+cmd.pycharm=Tool commandlet for Pycharm (IDE)
+cmd.pycharm.detail=Pycharm is a popular Integrated Development Environment for Python developed by JetBrains. Detailed documentation can be found at https://www.jetbrains.com/pycharm/documentation/
 cmd.python=Tool commandlet for Python.
 cmd.python.detail=Python is an object-oriented programming language, comparable to Perl, Ruby, Scheme, or Java. Detailed documentation can be found at https://www.python.org/doc/
 cmd.quarkus=Tool commandlet for Quarkus (framework for cloud-native apps).
