+++ cli/src/main/resources/nls/Help_de.properties
@@ -46,7 +46,7 @@ cmd.install-plugin=Installiert die selektierte Erweiterung für das selektierte 
 cmd.install-plugin.detail=Erweiterung können nur für Werkzeuge installiert werden, die diese unterstützen. Erweiterungen müssen in den Projekteinstellungen definiert sein um zur Installation verfügbar zu sein. Falls eine gewünschte Erweiterung nicht verfügbar ist, bitte den ide-admin kontaktieren.
 cmd.install.detail=Die Liste an zur Installation verfügbaren Werkzeugen können mit dem Befehl "ide help" ausgegeben werden.
 cmd.intellij=Werkzeug Kommando für Intellij (IDE)
-cmd.intellij.detail=IntelliJ ist eine beliebte Java-Entwicklungsumgebung, von JetBrains entwickelt. Detaillierte Dokumentation ist zu finden unter https://www.jetbrains.com/idea/documentation/
+cmd.intellij.detail=IntelliJ ist eine beliebte Java-Entwicklungsumgebung, die von JetBrains entwickelt wurde. Detaillierte Dokumentation ist zu finden unter https://www.jetbrains.com/idea/documentation/
 cmd.jasypt=Werkzeug Kommando für Jasypt.
 cmd.jasypt.detail=Jasypt ist eine Bibliothek zum Verschlüsseln und Entschlüsseln sensibler Daten in Java-Anwendungen. Zusätzliche Jasypt Optionen wie "algorithm" können mittels der Variable JASYPT_OPTS konfiguriert werden. Detaillierte Dokumentation ist zu finden unter http://www.jasypt.org/
 cmd.jasypt.val.command=Mode (encrypt | decrypt)
@@ -78,6 +78,8 @@ cmd.oc=Werkzeug Kommando für Openshift CLI (Kubernetes Management Tool).
 cmd.oc.detail=OpenShift CLI (oc) ist ein Befehlszeilentool zur Verwaltung von OpenShift-Clustern. Detaillierte Dokumentation ist zu finden unter https://docs.openshift.com/
 cmd.pgadmin=Werkzeug Kommando für pgAdmin.
 cmd.pgadmin.detail=pgAdmin ist ein Open-Source-Verwaltungswerkzeug for PostgreSQL und verwandte relationale Datenbanken. Detaillierte Dokumentation ist zu finden unter https://www.pgadmin.org/docs/pgadmin4/latest/index.html
+cmd.pycharm=Werkzeug Kommando für Pycharm (IDE)
+cmd.pycharm.detail=Pycharm ist eine beliebte Python-Entwicklungsumgebung, die von JetBrains entwickelt wurde. Detaillierte Dokumentation ist zu finden unter https://www.jetbrains.com/pycharm/documentation/
 cmd.python=Werkzeug Kommando für Python.
 cmd.python.detail=Python ist eine objektorientierte Programmiersprache, vergleichbar mit Perl, Ruby, Scheme oder Java. Detaillierte Dokumentation ist zu finden unter https://www.python.org/doc/
 cmd.quarkus=Werkzeug Kommando für Quarkus (Framework für Cloud-native Anwendungen).
