+++ cli/src/test/resources/ide-projects/pycharm/repository/pycharm/pycharm/default/windows/bin/pycharm64.exe
@@ -1,0 +1,11 @@
+#!/bin/bash
+cd "$(dirname "$0")"
+echo "pycharm windows $*" > pycharmtest
+
+if [ "${1}" == "installPlugins" ]; then
+  echo "installed plugin: PluginNode{id=${@:2}}"
+fi
+
+if [ -n "${3}" ]; then
+  echo "custom plugin repo url is: ${3}" > customRepoTest
+fi
